fix: imports of tests, fix blackbox test

+++ src/BSP/led_extractor.py
@@ -31,7 +31,7 @@ def get_led_roi(frame: np.array, leds: List[Led], board_orientation: BoardOrient
     radius = []
 
     for i in range(len(leds)):
-        radius.append(round(max(abs(led_centers_transformed[i][0] - led_radius_transformed[i][0]), abs(led_centers_transformed[i][1] - led_radius_transformed[i][1]))))
+        radius.append(int(round(max(abs(led_centers_transformed[i][0] - led_radius_transformed[i][0]), abs(led_centers_transformed[i][1] - led_radius_transformed[i][1])))))
 
     #radius = list(map(lambda led: round(led.radius * max(scale_x, scale_y)), leds))
     led_rois: List[np.array] = _led_by_circle_coordinates(frame, led_centers_transformed.astype(int), radius)
